+++ CHANGES.rst
@@ -13,6 +13,8 @@ Unreleased
 
 - Added support for cluster import jobs of type `file`.
 
+- Added ``idp`` to ``croud me``.
+
 1.3.0 - 2023/03/14
 ==================